requirements.txt
streamlit
requests
beautifulsoup4
pandas
lxml
sqlalchemy
tqdm
python-dotenv
google-genai